requirements.txt
Flask==0.10.1
Glances==1.6
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.3
SOAPpy==0.12.0
Werkzeug==0.9.6
argparse==1.2.1
chardet==2.0.1
configobj==4.7.2
decorator==3.3.3
fpconst==0.7.2
iotop==0.4.4
ipython==2.2.0
itsdangerous==0.24
medusa==0.5.4
meld3==0.6.5
mercurial==2.2.2
pexpect==2.4
pony==0.5.1
psutil==0.7.1
py-oauth2==0.0.10
python-apt==0.8.8.2
python-debian==0.1.21
python-debianbts==1.11
python-jenkins==0.3.4
reportbug==6.4.4
requests==2.4.1
sh==1.09
simplegeneric==0.8.1
six==1.8.0
supervisor==3.0a8
virtualenv==1.11.6
wsgiref==0.1.2
